Fixing build

+++ cpp-utils/test/codecs_tests.cpp
@@ -11,7 +11,7 @@ void test_vbe() {
     char buf[16];
     unsigned long x = 123123;
     char* p = vbe_encode(buf, x);
-    assert(p - buf == vbe_encoded_size(x));
+    assert((size_t)(p - buf) == vbe_encoded_size(x));
     unsigned long y;
     assert(vbe_decode(buf, y) == p);
     assert(y == x);
